Create day0_knn.py
In this Python code file, I demonstrate what I was learning today, how to implement KNN in Python with holdout validation (training and testing on two 50% splits of your dataset).  This is a pretty cool technique and can be pretty useful when you're trying predict, for example, a home price using attributes about the home.   Can be applied across a large number of domains.